refactor v1 trace arg helpers
